+++ tests/functional/test_issue_371/main.tf
@@ -1,0 +1,5 @@
+provider "aws" {
+  region     = "eu-west-3"
+  access_key = "AKIAIOSFODNN7EXAMPLE"
+  secret_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
+}